% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{atc}
\alias{atc}
\title{Index of ATC Codes}
\format{
An index of ATC codes, including compound names.
The following columns are included:
\itemize{
\item \code{ATC.Code} The ATC code.
\item \code{Name} Name of the compound corresponding to a given ATC.
}
}
\usage{
atc
}
\description{
Index of ATC Codes
}
\examples{
data("atc", package = "SCREAM")
}
\keyword{datasets}
