% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckd_epi.R
\name{ckd_epi}
\alias{ckd_epi}
\title{CKD-EPI Equation}
\usage{
ckd_epi(creatinine, age, female)
}
\arguments{
\item{creatinine}{Serum creatinine (in μmol/L);}

\item{age}{Age;}

\item{female}{Gender (with females identified by \code{female = 1}).}
}
\value{
Estimated eGFR, in mL/min/1.73m2.
}
\description{
CKD-EPI Equation
}
\note{
Compare e.g. with \url{https://www.kidney.org/professionals/kdoqi/gfr_calculator}.
}
\examples{
ckd_epi(creatinine = 50, age = 40, female = 0)
ckd_epi(creatinine = 50, age = 40, female = 1)
}
