% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{drug}
\alias{drug}
\title{Example Dataset with Drug Utilisation Codes}
\format{
A data frame with drug utilisation information for 3,000 subjects with the following columns:
\itemize{
\item \code{id} Subject ID.
\item \code{date} Date corresponding to each code.
\item \code{code} ATC code.
\item \code{npacks} Number of drug packages purchased at this given date.
}
}
\usage{
drug
}
\description{
Example Dataset with Drug Utilisation Codes
}
\examples{
data("drug", package = "SCREAM")
}
\keyword{datasets}
