% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd10}
\alias{icd10}
\title{Example of a dataset with ICD-10 codes}
\format{
A list with two simulated data frames of ICD-10 codes for 3,000 subjects, one with hospitalisation codes, and one with inpatient and outpatient claims.
Each data set includes the following variables:
\itemize{
\item \code{id} Subject ID.
\item \code{date} Date corresponding to each code.
\item \code{index_date} Index date (used e.g. to calculate multimorbidity at that specific point in time).
\item \code{code} ICD-10 code.
}
}
\usage{
icd10
}
\description{
Example of a dataset with ICD-10 codes
}
\examples{
data("icd10", package = "SCREAM")
}
\keyword{datasets}
