% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_date.R
\name{interpolate_date}
\alias{interpolate_date}
\title{Interpolate kidney function via a linear mixed effects model}
\usage{
interpolate_date(fit, egfr, id, origin, time_scale)
}
\arguments{
\item{fit}{A mixed model fitted using \code{\link[lme4:lmer]{lme4::lmer()}}.}

\item{egfr}{Interpolation identifies the date at which \code{egfr} is reached.
Can be a vector or a single value.}

\item{id}{Name of the ID variable used to identify each study subject.}

\item{origin}{Origin of the time scale.
Can be a vector of dates or a single value.}

\item{time_scale}{Time scale used for the calculations; can be either \code{"years"} or \code{"days"}.}
}
\value{
A data frame with a row per subject and a column per \code{egfr} value.
}
\description{
Interpolate kidney function via a linear mixed effects model
}
