% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_survival_data}
\alias{sample_survival_data}
\title{Example Dataset of survival outcome}
\format{
A data frame of 1,000 subjects with the following columns:
\itemize{
\item \code{lopnr} subject unique ID.
\item \code{index_date} index date.
\item \code{tstart_as_Date} or \code{tstart} index date with class \code{Date} or type \code{numeric}.
\item \code{tstop_as_Date} or \code{tstop} outcome/censoring date with class \code{Date} or type \code{numeric}.
\item \code{outcome_event} indicator of the survival outcome.
\item \code{null_weight} weight for each included patients.
}
}
\usage{
sample_survival_data
}
\description{
Example Dataset of survival outcome
}
\examples{
data("sample_survival_data", package = "SCREAM")
}
\keyword{datasets}
