% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scream-package.R
\docType{package}
\name{SCREAM}
\alias{SCREAM}
\title{Utility functions for data analysis in SCREAM}
\description{
Utility function for data analysis using data from the Stockholm
CREAtinine Measurements project.
}
\author{
Alessandro Gasparini (alessandro.gasparini@ki.se)
}
