% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_outcome_summary.R
\name{survival_outcome_summary}
\alias{survival_outcome_summary}
\title{Summarise the basic information of a survival outcome}
\usage{
survival_outcome_summary(
  outcome,
  tstart,
  tstop,
  exposure,
  weight,
  dataset,
  digit = 2
)
}
\arguments{
\item{outcome}{survival outcome of interest.}

\item{tstart}{index date with class \code{"Date"} or numeric value \code{0}.}

\item{tstop}{outcome occurrence date or censoring date with class \code{"Date"} or
numeric value of days from index date to outcome occurrence date or censoring date.}

\item{exposure}{the summary of the survival outcome is usually displayed in each exposure
group, if the summary of the survival outcome for all included patients should displayed,
set \code{exposure = "all"}.}

\item{weight}{weight for each included patient, in crude cohort dataset, it is equal to \code{1},
in propensity-score weighted cohort dataset, it is equal to \eqn{\frac{1}{PS}} for exposed group,
and \eqn{\frac{1}{1-PS}} for noexposed group.}

\item{dataset}{dataset of the cohort study, in a time-fixed format of one patient one row.}

\item{digit}{number of digits after the decimal point}
}
\value{
a dataframe showing the summary of the survival outcome.
}
\description{
\code{survival_outcome_summary} is a function used to produce a summary
of a survival outcome in a cohort study, including number of patients,
follow-up time (median \link{IQR}, years), number of events, person years,
rate with 95\% confidence interval (per 1000 person years).
}
\examples{
data("sample_survival_data", package = "SCREAM")
survival_outcome_summary("outcome_event",
    "tstart_as_Date",
    "tstop_as_Date",
    "all",
    "null_weight",
    sample_survival_data,
    digit = 2)
}
