% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{icd10}
\alias{icd10}
\title{Example of a dataset with ICD-10 codes}
\format{
A data frame with 123,772 rows and 7 variables, for 5,000 subjects. Included variables are:
\itemize{
\item \code{id} Subject ID.
\item \code{date} Date corresponding to each code.
\item \code{index_date} Index date (used e.g. to calculate multimorbidity at that specific point in time).
\item \code{code} ICD-10 code.
\item \code{origin} Origin of each ICD-10 code, with \code{origin = 1} denoting hospital care, \code{origin = 2} denoting secondary care, and \code{origin = 3} denoting primary care.
\item \code{n} Sequential number for each observation.
\item \code{N} Total number of observations per subject.
}
}
\usage{
icd10
}
\description{
Example of a dataset with ICD-10 codes
}
\examples{
data("icd10", package = "SCREAM")
}
\keyword{datasets}
