% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_long_egfr.R
\name{simulate_long_egfr}
\alias{simulate_long_egfr}
\title{Simulate longitudinal eGFR data}
\usage{
simulate_long_egfr(
  n,
  id = "lopnr",
  fixed_intercept = 100,
  fixed_slope = -1,
  B = matrix(data = c(100, 10, 10, 5), nrow = 2, ncol = 2),
  sigma = 10,
  lambda = 1,
  size = 1,
  prob = 0.1
)
}
\arguments{
\item{n}{Number of observations.}

\item{id}{Name of the ID variable that will be used throughout.}

\item{fixed_intercept}{Fixed intercept.}

\item{fixed_slope}{Fixed slope.}

\item{B}{Variance-covariance matrix of the random effects.}

\item{sigma}{Standard deviation of the residual error term.}

\item{lambda}{Parameter \eqn{lambda} of the baseline hazard for the exponential survival model.}

\item{size}{Parameter \code{'size'} for the negative binomial distribution; see \code{\link[stats:NegBinomial]{stats::dnbinom()}}.}

\item{prob}{Parameter \code{'prob'} for the negative binomial distribution; see \code{\link[stats:NegBinomial]{stats::dnbinom()}}.}
}
\value{
A data frame with simulate data for \code{n} subjects.
}
\description{
Simulate longitudinal eGFR data
}
\examples{

df <- simulate_long_egfr(n = 3)
df
}
